Delete financial data box component and add back to project display
