+++ app/projects/[slug]/project-display.tsx
@@ -3,11 +3,10 @@ import { DonateBox } from '@/components/donate-box'
 import { Col } from '@/components/layout/col'
 import { Row } from '@/components/layout/row'
 import { ProgressBar } from '@/components/progress-bar'
-import { SignInButton } from '@/components/sign-in-button'
 import { BidAndProfile, BidAndProject } from '@/db/bid'
 import { CommentAndProfile } from '@/db/comment'
 import { Profile } from '@/db/profile'
-import { FullProject } from '@/db/project'
+import { FullProject, Project } from '@/db/project'
 import { Cause, SimpleCause } from '@/db/cause'
 import { TxnAndProfiles, TxnAndProject } from '@/db/txn'
 import {
@@ -35,8 +34,8 @@ import { UserAvatarAndBadge } from '@/components/user-link'
 import { Tooltip } from '@/components/tooltip'
 import { EnvelopeIcon } from '@heroicons/react/20/solid'
 import { ViewerActionPanel } from './viewer-action-panel'
-import { FinancialDataBox, ProjectTypeDisplay } from './financial-data-box'
 import { differenceInDays, differenceInHours } from 'date-fns'
+import { toSentenceCase } from '@/utils/formatting'
 
 export function ProjectDisplay(props: {
   project: FullProject
@@ -183,37 +182,38 @@ export function ProjectDisplay(props: {
             )}
           </Row>
         </div>
-        {/* <ProjectData
-            project={project}
-            raised={amountRaised}
-            valuation={valuation}
-            minimum={minIncludingAmm}
-          />
-        {project.stage !== 'proposal' &&
-          project.type === 'cert' &&
-          !!project.amm_shares && (
-            <CertValuationChart
-              tradePoints={tradePoints}
-              ammTxns={projectTxns.filter(
-                (txn) => txn.to_id === project.id || txn.from_id === project.id
-              )}
-              ammId={project.id}
-              size="lg"
+        <Col className="mb-4 gap-6 border-y border-gray-200 py-6">
+          <ProjectTypeDisplay type={project.type} stage={project.stage} />
+          <div>
+            {(project.stage === 'proposal' ||
+              (project.stage === 'active' && project.type === 'grant')) && (
+              <ProgressBar
+                amountRaised={amountRaised}
+                minFunding={minIncludingAmm}
+                fundingGoal={project.funding_goal}
+              />
+            )}
+            <ProjectData
+              minimum={minIncludingAmm}
+              valuation={valuation}
+              project={project}
+              raised={amountRaised}
             />
-          )}
-        {(project.stage === 'proposal' ||
-          (project.stage === 'active' && project.type === 'grant')) && (
-          <ProgressBar
-            amountRaised={amountRaised}
-            minFunding={minIncludingAmm}
-            fundingGoal={
-              project.type === 'cert' ? minIncludingAmm : project.funding_goal
-            }
-          />
-        )}
-        {userProfile &&
-          project.type === 'cert' &&
-          project.stage === 'active' && (
+          </div>
+          {!['draft', 'proposal'].includes(project.stage) &&
+            project.type === 'cert' &&
+            !!project.amm_shares && (
+              <CertValuationChart
+                tradePoints={tradePoints}
+                ammTxns={projectTxns.filter(
+                  (txn) =>
+                    txn.to_id === project.id || txn.from_id === project.id
+                )}
+                ammId={project.id}
+                size="lg"
+              />
+            )}
+          {project.type === 'cert' && project.stage === 'active' && (
             <Trade
               ammTxns={
                 !!project.amm_shares
@@ -226,59 +226,45 @@ export function ProjectDisplay(props: {
               projectId={project.id}
               userSpendableFunds={userSpendableFunds}
               userSellableShares={userSellableShares}
+              signedIn={!!userProfile}
             />
           )}
-        {userProfile &&
-          userProfile.id !== project.creator &&
-          project.type === 'cert' &&
-          project.stage === 'proposal' && (
-            <AssuranceBuyBox
-              project={project}
-              minValuation={valuation}
-              offerSizePortion={
-                (activeAuction
-                  ? minIncludingAmm
-                  : minIncludingAmm - amountRaised) / valuation
-              }
-              maxBuy={userSpendableFunds}
-              activeAuction={activeAuction}
-            />
-          )}
-        {userProfile &&
-          userProfile.id !== project.creator &&
-          project.type === 'grant' &&
-          pendingProjectTransfers.length === 0 &&
-          (project.stage === 'proposal' || project.stage === 'active') && (
-            <>
-              {amountRaised < project.funding_goal ? (
-                <DonateBox
-                  project={project}
-                  profile={userProfile}
-                  maxDonation={userSpendableFunds}
-                  setCommentPrompt={setSpecialCommentPrompt}
-                />
-              ) : (
-                <span className="mx-auto mb-5 text-sm italic text-gray-500">
-                  Fully funded and not currently accepting donations.
-                </span>
-              )}
-            </>
-          )}
-        {!userProfile && (
-          <SignInButton
-            buttonText="Sign in to contribute"
-            className="mx-auto my-4"
-          />
-        )} */}
-        <FinancialDataBox
-          project={project}
-          userTxns={userTxns}
-          userBids={userBids}
-          projectBids={projectBids}
-          projectTxns={projectTxns}
-          userProfile={userProfile}
-          prizeCause={prizeCause}
-        />
+          {userProfile?.id !== project.creator &&
+            project.type === 'cert' &&
+            project.stage === 'proposal' && (
+              <AssuranceBuyBox
+                project={project}
+                minValuation={valuation}
+                offerSizePortion={
+                  (activeAuction
+                    ? minIncludingAmm
+                    : minIncludingAmm - amountRaised) / valuation
+                }
+                maxBuy={userSpendableFunds}
+                activeAuction={activeAuction}
+                signedIn={!!userProfile}
+              />
+            )}
+          {userProfile?.id !== project.creator &&
+            project.type === 'grant' &&
+            pendingProjectTransfers.length === 0 &&
+            (project.stage === 'proposal' || project.stage === 'active') && (
+              <>
+                {amountRaised < project.funding_goal ? (
+                  <DonateBox
+                    project={project}
+                    profile={userProfile}
+                    maxDonation={userSpendableFunds}
+                    setCommentPrompt={setSpecialCommentPrompt}
+                  />
+                ) : (
+                  <span className="mx-auto mb-5 text-sm italic text-gray-500">
+                    Fully funded and not currently accepting donations.
+                  </span>
+                )}
+              </>
+            )}
+        </Col>
         {project.description && (
           <RichContent content={project.description} className="px-3 text-sm" />
         )}
@@ -325,4 +311,33 @@ export function scrollToComments(router: any) {
     block: 'start',
     inline: 'start',
   })
+}
+
+export function ProjectTypeDisplay(props: {
+  type: Project['type']
+  stage: Project['stage']
+}) {
+  const { type, stage } = props
+  return (
+    <Row className="justify-between">
+      <SmallDescriptionPoint label="stage" value={toSentenceCase(stage)} />
+      <SmallDescriptionPoint label="type" value={typeToLabelMap[type]} />
+    </Row>
+  )
+}
+
+function SmallDescriptionPoint(props: { label: string; value: string }) {
+  const { label, value } = props
+  return (
+    <div className="px-4 sm:px-0">
+      <dt className="text-sm leading-6 text-gray-700">{label}</dt>
+      <dd className="text-sm font-medium leading-6 text-gray-900">{value}</dd>
+    </div>
+  )
+}
+
+const typeToLabelMap = {
+  cert: 'Impact certificate',
+  grant: 'Grant',
+  dummy: 'Uninteractable stand in',
 }
